refactor: Move status/readiness patching into own funcs
